+++ RetailCoder.VBE/UI/Inspections/CodeInspectionsWindow.resx
