+++ RetailCoder.VBE/UI/Inspections/InspectionResultsControl.xaml
@@ -1,10 +1,10 @@
-﻿<UserControl x:Class="Rubberduck.UI.CodeInspections.InspectionResultsControl"
+﻿<UserControl x:Class="Rubberduck.UI.Inspections.InspectionResultsControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:inspections="clr-namespace:Rubberduck.Inspections"
-             xmlns:codeInspections="clr-namespace:Rubberduck.UI.CodeInspections"
+             xmlns:codeInspections="clr-namespace:Rubberduck.UI.Inspections"
              xmlns:controls="clr-namespace:Rubberduck.UI.Controls"
              xmlns:componentModel="clr-namespace:System.ComponentModel;assembly=WindowsBase"
              xmlns:themes="clr-namespace:Microsoft.Windows.Themes;assembly=PresentationFramework.Aero"
