+++ RetailCoder.VBE/UI/Inspections/InspectionSeverityImageSourceConverter.cs
@@ -8,7 +8,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using Rubberduck.Inspections;
 
-namespace Rubberduck.UI.CodeInspections
+namespace Rubberduck.UI.Inspections
 {
     public class InspectionSeverityImageSourceConverter : IValueConverter
     {
